Passing Object to Method
The given Java program demonstrates how to modify an object passed as a parameter to a method. In this case, the modifyCar method is used to change the model of a Car object.

Here's a breakdown of the code:

Car class: Defines a simple Car class with a model attribute.

CarProcessor class:

modifyCar method: Takes a Car object and a new model as parameters and modifies the model attribute of the Car object.
main method: Creates a Car object with the model "Toyota", prints the model before modification, calls the modifyCar method to change the model to "Honda", and then prints the model after modification.