Object Stream Demonstration
This Example Demonstrate Object Steam Writing to file. 
Here we need use Serializable Interface save the object. 

In Java, the Serializable interface serves as a marker interface, which means it doesn't have any methods to implement. Instead, its presence in a class's declaration indicates to the Java Virtual Machine (JVM) that instances of that class can be serialized. Serialization is the process of converting an object into a stream of bytes so that it can be stored in memory, in a file, or transmitted across a network.