Create  InitStaticBlockEx .java
Init and Static Block Examples

Instance Initialization Block (IIB):

init:- An instance initialization block is a block of code within a class that is executed each time an instance of the class is created.
It is used to initialize instance variables of the class.
Instance initialization blocks are defined without any keyword, simply within curly braces {}.

Static Initialization Block (SIB):

A static initialization block is a block of code within a class that is executed once when the class is loaded into memory by the class loader.
It is used to initialize static variables of the class or perform any one-time initialization tasks for the class.
Static initialization blocks are defined using the static keyword.

Initialization blocks are useful when you need to perform certain tasks before object creation (in the case of instance initialization blocks) or before class initialization (in the case of static initialization blocks). They provide a way to organize and centralize initialization logic within a class.