Program to Demonstrate Mark Invalidate
To demonstrate the mark getting invalidated, you can reduce the buffer size of the BufferedInputStream to a value smaller than the read limit you set with mark(int readlimit). This will force the stream to read data directly from the underlying FileInputStream more frequently, potentially causing the mark to be invalidated when you try to read beyond the read limit.   I've created the BufferedInputStream with a buffer size of 2 bytes explicitly (BufferedInputStream(fis, 2)). This ensures that the buffer size is smaller than the read limit of 3 characters. As a result, when you try to read more than 3 characters after setting the mark, the mark will be invalidated because the buffer will be exhausted, and the stream will need to read data from the underlying FileInputStream again.