Arras Of Object Soring Example
Arrays.sort(...): This method is used to sort arrays in Java. It accepts four parameters:
The array to be sorted.
The starting index of the range to be sorted.
The ending index of the range to be sorted (exclusive).
The comparator to determine the order of elements.
s1: This is the array of Student objects that you want to sort.
0: This is the starting index of the range to be sorted. In this case, it starts from index 0.
4: This is the ending index of the range to be sorted (exclusive). It means the sorting will be applied to elements from index 0 to index 3 of the array s1.
Comparator.comparing(Student::getPage): This part specifies the primary sorting criteria. It tells the sort method to compare Student objects based on their page attribute. So, it sorts the students based on their page numbers in ascending order.
.thenComparing(Student::getName): This part specifies the secondary sorting criteria. It tells the sort method that if two students have the same page value, then compare them based on their name attribute. This ensures that if two students have the same page number, they will be sorted based on their names in alphabetical order.
Overall, this line of code sorts the array s1 of Student objects based on the page attribute first, and if there are any students with the same page value, it sorts them based on their name attribute.