Create BubbleSort.java
In this example, we have used an array of integers to store the elements, and the bubble sort algorithm is implemented inside two for loops. The program asks the user to enter the size of the array and the elements of the array, then displays the sorted array at the end.