Create SelectionSort.java
In this example, we have used the selection sort algorithm to sort an array of integers. The program asks the user to enter the size of the array and the elements, then uses a double loop to find the minimum element in the unsorted portion of the array and places it in the correct position. This operation is repeated until the entire array is sorted.