Create BinarySearch.java
In this example, we have used the binary search algorithm to find an element in a sorted array. The program asks the user to enter the size of the sorted array, the elements of the array and the element to be searched for. It then performs the binary  search to find the element and displays its position if it is found, or indicates if it has not been found.