Create SequentialSearch.java
In this example, we have used the sequential search algorithm to find an element in an array. The program asks the user to enter the size of the array, the elements of the array and the element to be searched for. It then performs the sequential search to find the element and displays its position if it is found, or indicates if it has not been found. Sequential search is simple but can be less efficient than other methods for large arrays, as it traverses all the
elements in the array.